+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Partials `site_header` and `site_footer` for easy overrides
+* A helper method `#site_title`
 
 ### Changed